@@ -3,25 +3,24 @@ using Application.EventSourcing;
 using Application.Infrastructure;
 using Scheduling.Domain.Infrastructure;
 
-namespace Scheduling.Test
+namespace Scheduling.Test;
+
+public class FakeAggregateStore : IAggregateStore
 {
-    public class FakeAggregateStore : IAggregateStore
+    private readonly AggregateRoot _aggregateRoot;
+
+    public FakeAggregateStore(AggregateRoot aggregateRoot)
     {
-        private readonly AggregateRoot _aggregateRoot;
+        _aggregateRoot = aggregateRoot;
+    }
 
-        public FakeAggregateStore(AggregateRoot aggregateRoot)
-        {
-            _aggregateRoot = aggregateRoot;
-        }
+    public Task Save<T>(T aggregate) where T : AggregateRoot
+    {
+        return Task.CompletedTask;
+    }
 
-        public Task Save<T>(T aggregate) where T : AggregateRoot
-        {
-            return Task.CompletedTask;
-        }
-
-        public Task<T> Load<T>(string aggregateId) where T : AggregateRoot
-        {
-            return Task.FromResult((T) _aggregateRoot);
-        }
+    public Task<T> Load<T>(string aggregateId) where T : AggregateRoot
+    {
+        return Task.FromResult((T) _aggregateRoot);
     }
 }
