@@ -1,17 +1,8 @@
 using Application.EventSourcing;
 
-namespace Application.Domain.WriteModel.Commands
-{
-    public class Book : Command
-    {
-        public string Id { get; }
+namespace Application.Domain.WriteModel.Commands;
 
-        public string PatientId { get; }
-
-        public Book(string id, string patientId)
-        {
-            Id = id;
-            PatientId = patientId;
-        }
-    }
-}
+public record Book(
+    string Id,
+    string PatientId
+) : ICommand;
