@@ -2,52 +2,51 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Application.Infrastructure
+namespace Application.Infrastructure;
+
+public abstract class AggregateRoot
 {
-    public abstract class AggregateRoot
+    readonly Dictionary<Type, Action<object>> _handlers = new();
+
+    readonly List<object> _changes = new();
+
+    public string Id { get; set; } = default!;
+
+    public int Version { get; protected set; } = -1;
+
+    protected void Register<T>(Action<T> when) =>
+        _handlers.Add(typeof(T), e => when((T) e));
+
+    protected void Raise(object e)
     {
-        readonly Dictionary<Type, Action<object>> _handlers = new Dictionary<Type, Action<object>>();
+        _handlers.TryGetValue(e.GetType(), out var handler);
 
-        readonly List<object> _changes = new List<object>();
-
-        public string Id { get; set; }
-
-        public int Version { get; protected set; } = -1;
-
-        protected void Register<T>(Action<T> when) =>
-            _handlers.Add(typeof(T), e => when((T) e));
-
-        protected void Raise(object e)
+        if (handler == null)
         {
-            _handlers.TryGetValue(e.GetType(), out var handler);
-
-            if (handler == null)
-            {
-                throw new NoHandlerRegisteredException(e.GetType());
-            }
-
-            handler(e);
-            _changes.Add(e);
+            throw new NoHandlerRegisteredException(e.GetType());
         }
 
-        public IEnumerable<object> GetChanges() => _changes.AsEnumerable();
-
-        public void Load(IEnumerable<object> history)
-        {
-            foreach (var e in history)
-            {
-                Raise(e);
-                Version++;
-            }
-        }
-
-        public void ClearChanges() => _changes.Clear();
+        handler(e);
+        _changes.Add(e);
     }
 
-    public class NoHandlerRegisteredException : Exception
+    public IEnumerable<object> GetChanges() => _changes.AsEnumerable();
+
+    public void Load(IEnumerable<object> history)
     {
-        public NoHandlerRegisteredException(Type type) : base($"No handler registered for {type.FullName}")
+        foreach (var e in history)
         {
+            Raise(e);
+            Version++;
         }
+    }
+
+    public void ClearChanges() => _changes.Clear();
+}
+
+public class NoHandlerRegisteredException : Exception
+{
+    public NoHandlerRegisteredException(Type type) : base($"No handler registered for {type.FullName}")
+    {
     }
 }
