@@ -1,9 +1,8 @@
 using System.Collections.Generic;
 
-namespace Application.EventSourcing
+namespace Application.EventSourcing;
+
+public abstract class AggregateState
 {
-    public abstract class AggregateState
-    {
-        public abstract void Apply(List<Event> events);
-    }
+    public abstract void Apply(List<IEvent> events);
 }
