@@ -3,39 +3,44 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Application.Infrastructure.Projections
+namespace Application.Infrastructure.Projections;
+
+public class Projection
 {
-    public class Projection
+    readonly List<EventHandler> _handlers = new();
+
+    protected void When<T>(Func<T, Task> when)
     {
-        readonly List<EventHandler> _handlers = new List<EventHandler>();
+        _handlers.Add(new EventHandler(typeof(T), async e => await when((T)e)));
+    }
 
-        protected void When<T>(Func<T, Task> when)
+    public async Task Handle(Type eventType, object e)
+    {
+        var handlers = _handlers
+            .Where(h => h.EventType == eventType)
+            .ToList();
+
+        foreach (var handler in handlers)
         {
-            _handlers.Add(new EventHandler { EventType = typeof(T), Handler = async e => await when((T)e) });
+            await handler.Handler(e);
         }
+    }
 
-        public async Task Handle(Type eventType, object e)
+    public bool CanHandle(Type eventType)
+    {
+        return _handlers.Any(h => h.EventType == eventType);
+    }
+
+    public class EventHandler
+    {
+        public Type EventType { get; }
+        
+        public Func<object, Task> Handler { get; }
+        
+        public EventHandler(Type eventType, Func<object, Task> handler)
         {
-            var handlers = _handlers
-                .Where(h => h.EventType == eventType)
-                .ToList();
-
-            foreach (var handler in handlers)
-            {
-                await handler.Handler(e);
-            }
-        }
-
-        public bool CanHandle(Type eventType)
-        {
-            return _handlers.Any(h => h.EventType == eventType);
-        }
-
-        public class EventHandler
-        {
-            public Type EventType { get; set; }
-
-            public Func<object, Task> Handler { get; set; }
+            EventType = eventType;
+            Handler = handler;
         }
     }
 }
