@@ -1,27 +1,26 @@
 using System;
 
-namespace Application.Infrastructure.Projections
+namespace Application.Infrastructure.Projections;
+
+public class StreamName
 {
-    public class StreamName
-    {
-        string Value { get; }
-        const string AllStreamName = "$all";
+    string Value { get; }
+    const string AllStreamName = "$all";
 
-        StreamName(string value) => Value = value;
+    StreamName(string value) => Value = value;
 
-        public static StreamName AllStream => new StreamName(AllStreamName);
+    public static StreamName AllStream => new StreamName(AllStreamName);
 
-        public static StreamName For<T>(Guid id)
-            => new StreamName($"{typeof(T).Name}-{id:N}");
+    public static StreamName For<T>(Guid id)
+        => new StreamName($"{typeof(T).Name}-{id:N}");
 
-        public static StreamName Custom(string streamName)
-            => new StreamName(streamName);
+    public static StreamName Custom(string streamName)
+        => new StreamName(streamName);
 
-        public bool IsAllStream => Value.Equals(AllStreamName);
+    public bool IsAllStream => Value.Equals(AllStreamName);
 
-        public override string ToString() => Value ?? "";
+    public override string ToString() => Value ?? "";
 
-        public static implicit operator string(StreamName self)
-            => self.ToString();
-    }
+    public static implicit operator string(StreamName self)
+        => self.ToString();
 }
